ATR-23: Enumerable refactoring
